+++ src/main/java/com/fanfan/service/impl/EmployeeServiceImpl.java
@@ -46,7 +46,7 @@ public class EmployeeServiceImpl implements EmployeeService {
     public PageBean page(int page, int pageSize, String name) {
         PageHelper.startPage(page, pageSize);
         Page<Employee> p = employeeMapper.findPage(name);
-        return new PageBean(p.getTotal(), p.getResult());
+        return new PageBean((int) p.getTotal(), p.getResult());
     }
 
     /**
